moved hero component to its folder
